# Optimize `TextHelper.highlight` for large inputs