Fixes #47337.

Previous regex used lookahead, so it was `O(n^2)` of the string size - for each detected phrase it looked forward (possibly for the rest of the string till the end) to see if it is not in the middle of the html tag.

Current approach linearly scans the text.

For the larger text (`highlight("foo bar 1234\n" * 10000, /foo|bar|1234/)`), previously it ran 30s, now - ~1s.

For smaller texts, performance is basically the same:
```ruby
Benchmark.ips do |x|
  x.report("original") do
    TestController.helpers.highlight('You searched for: rails', /for|rails/)
  end

  x.report("optimized") do
    TestController.helpers.highlight_optimized('You searched for: rails', /for|rails/)
  end

  x.compare!
end
```

```
Warming up --------------------------------------
            original   330.000  i/100ms
           optimized   357.000  i/100ms
Calculating -------------------------------------
            original      4.759k (±22.4%) i/s -     22.770k in   5.085356s
           optimized      3.559k (±18.7%) i/s -     17.493k in   5.106094s

Comparison:
            original:     4759.0 i/s
           optimized:     3559.1 i/s - same-ish: difference falls within error
```