@matthewd 2023-02-09
Nice! This does seem like it's going to do a lot of work in a large tag-rich input even with few highlight matches, though... what do you think of using `MatchData#end(0)` & `String#index(/[<>]/, offset)` to implement the previous "scan forward from end of match" behaviour, but avoid the quadratic by remembering the last scanned index?