@fatkodima 2023-02-09
> but avoid the quadratic by remembering the last scanned index

Isn't it linear? 🤔 

I used the suggested approach and got this:
```ruby
def do_highlight_optimized(text, pattern, highlighter: nil, &block)
  bracket_index = nil
  bracket = nil

  text.gsub(/#{pattern}/i) do |match|
    matchdata = Regexp.last_match
    offset = matchdata.end(0)

    if bracket_index.nil? || offset > bracket_index
      bracket_index = text.index(/[<>]/, offset)
      bracket_index ||= text.size
      bracket = text[bracket_index]
    end

    if bracket == ">"
      match
    elsif block_given?
      yield match
    else
      highlighter.gsub('\1', match)
    end
  end
end
```

Measured a time it took to run:
```ruby
# frozen_string_literal: true

require "bundler/inline"

gemfile(true) do
  source "https://rubygems.org"

  git_source(:github) { |repo| "https://github.com/#{repo}.git" }

  # Activate the gem you are reporting the issue against.
  gem "rails", path: "~/Desktop/oss/rails" # "~> 7.0.0"
  gem "benchmark-ips"
end

require "action_controller/railtie"

class TestApp < Rails::Application
  config.root = __dir__
  config.hosts << "example.org"
  config.session_store :cookie_store, key: "cookie_store_key"
  secrets.secret_key_base = "secret_key_base"

  config.logger = Logger.new($stdout)
  Rails.logger  = config.logger
end

class TestController < ActionController::Base
end

puts "Started"
start = Process.clock_gettime(Process::CLOCK_MONOTONIC)
input = "<zoo>woo</zoo> baz\n" * 10000 + "foo"
TestController.helpers.highlight_optimized(input, /foo|bar|1234/)
delta = Process.clock_gettime(Process::CLOCK_MONOTONIC) - start
puts "Finished in #{delta.round(2)} seconds"
```

And got the same results for this version and version in the PR. Am I missing something?