@matthewd 2023-02-10
> Isn't it linear?

Sorry, my phrasing was unclear. I meant to describe the way in which my proposal avoided the _original_ quadratic, by way of contrast to the also-linear strategy in your change.

That was what I had in mind, thanks for trying it -- I was definitely expecting the many allocations from the extra calls to be a time sink, but I guess not!